moved some components files
