+++ app/profile/page.tsx
@@ -6,7 +6,7 @@ import { ThemeToggle } from "@/shared/ui/theme-toggle"
 import { UserMenu } from "@/widgets/user-menu/user-menu"
 import { NavLinks } from "@/components/nav-links"
 import { useEffect } from "react"
-import WithoutAuth from "@/components/WithoutAuth"
+import WithoutAuth from "@/features/auth/ui/WithoutAuth"
 import { StatisticsDashboard } from "../../components/statistics/statistics-dashboard"
 import ProfileSettings from "@/components/profile-settings"
 import Subscription from "@/components/subscription"
