+++ shared/ui/YandexIcon.tsx
